requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
SQLAlchemy==1.4.50
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
python-docx==0.8.11
streamlit==1.27.0
requests==2.31.0